requirements.txt
asgiref==3.8.1
black==24.4.2
boto3==1.34.108
botocore==1.34.108
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
flake8==7.0.0
gunicorn==22.0.0
httpie==3.2.3
idna==3.7
jmespath==1.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.1.19
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
s3transfer==0.10.1
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
